Update script to work with the new TX command
Previous solution of patching .tx/config does not work with the
new client, so the script had to be updated with new solution.
Now, translations are pulled into a sub-directory of cpython, and
then installed in the languages's repository.

Also using 'realpath' tool in ROOTDIR to retrieve absolute path,
which actually eases referencing the repository root directory.