Improve ci (#40)
* Use new Transifex CLI tool

The Python implementation named transifex-client uses Transifex's
APIv2, which is deprecated and will sunset in november 2022.

The new tool, called Transifex CLI, is written in Golang and is
APIv3-ready. So adjusting the CI and requirements.txt is required.

* Fix condition in the merge job

Having the proper condition is important to have the translation
update in the latest Python version (python-newest project), and
the condition wouldn't match the current repository.

Co-authored-by: Maciej Olko <maciej.olko@affirm.com>

* Update sphinx-intl required version

Last released sphinx-intl uses old transifex-client, which uses
the deprecated Transifex APIv2. There is a fix committed, but
it is unreleased. So using specific commit where new Transifex
CLI tool works.

* Update script to work with the new TX command

Previous solution of patching .tx/config does not work with the
new client, so the script had to be updated with new solution.
Now, translations are pulled into a sub-directory of cpython, and
then installed in the languages's repository.

Also using 'realpath' tool in ROOTDIR to retrieve absolute path,
which actually eases referencing the repository root directory.

* Update actions versions

One of the reason is that some of them use node.js 12, which is
now deprecated. Actions is displaying warnings about it.

v4 of setup-python is particular useful because it introduces
package caching, which reduces the time of workflow run.

* Disable pushing translations

Source files auto-update is now done elsewhere

Co-authored-by: Maciej Olko <maciej.olko@affirm.com>